Updated food
I updated food to make a training set of 75% and a test set of 25%.
Everytime that the createReviewArray is called the test and training set
returned will be redone. I also updated the createReviewerToReviewMap to
return a map for a test set and a training set. The keys are still the
reviewers and the values are still the review tuples.