POS
I wsa able to find a POS. To install it go here
http://textblob.readthedocs.org/en/latest/install.html

ex. "This is a sentence that I am going to try and tag. Lets see if it
works!"
[('This', 'DT'), ('is', 'VBZ'), ('a', 'DT'), ('sentence', 'NN'),
('that', 'IN'), ('I', 'PRP'), ('am', 'VBP'), ('going', 'VBG'), ('to',
'TO'), ('try', 'VB'), ('and', 'CC'), ('tag', 'VB'), ('Lets', 'NNS'),
('see', 'VBP'), ('if', 'IN'), ('it', 'PRP'), ('works', 'VBZ')]